WIP: add some comments and todos
